Move GSON and Retrofit support classes to single infrastructure package.
