+++ JavaModules/idispatch-client/src/main/java/net/pkhapps/idispatch/client/v3/infrastructure/ColorJsonTypeAdapter.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.client.v3.util.gson;
+package net.pkhapps.idispatch.client.v3.infrastructure;
 
 import com.google.gson.*;
 import net.pkhapps.idispatch.client.v3.util.Color;
@@ -8,7 +8,7 @@ import java.lang.reflect.Type;
 /**
  * Type adapter ({@link JsonSerializer} and {@link JsonDeserializer}) for {@link Color}.
  */
-public class ColorJsonTypeAdapter implements JsonSerializer<Color>, JsonDeserializer<Color> {
+class ColorJsonTypeAdapter implements JsonSerializer<Color>, JsonDeserializer<Color> {
 
     @Override
     public JsonElement serialize(Color color, Type type, JsonSerializationContext jsonSerializationContext) {
