+++ JavaModules/idispatch-client/src/main/java/net/pkhapps/idispatch/client/v3/infrastructure/PhoneNumberJsonTypeAdapter.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.client.v3.util.gson;
+package net.pkhapps.idispatch.client.v3.infrastructure;
 
 import com.google.gson.*;
 import net.pkhapps.idispatch.client.v3.util.PhoneNumber;
@@ -8,7 +8,7 @@ import java.lang.reflect.Type;
 /**
  * Type adapter ({@link JsonSerializer} and {@link JsonDeserializer}) for {@link PhoneNumber}.
  */
-public class PhoneNumberJsonTypeAdapter implements JsonSerializer<PhoneNumber>, JsonDeserializer<PhoneNumber> {
+class PhoneNumberJsonTypeAdapter implements JsonSerializer<PhoneNumber>, JsonDeserializer<PhoneNumber> {
 
     @Override
     public JsonElement serialize(PhoneNumber phoneNumber, Type type,
